+++ demisto_sdk/commands/init/initiator.py
@@ -16,7 +16,7 @@ from demisto_sdk.commands.common.constants import (
     INDICATOR_TYPES_DIR, INTEGRATION_CATEGORIES, INTEGRATIONS_DIR, LAYOUTS_DIR,
     MARKETPLACE_LIVE_DISCUSSIONS, PACK_INITIAL_VERSION, PACK_SUPPORT_OPTIONS,
     PLAYBOOKS_DIR, REPORTS_DIR, SCRIPTS_DIR, TEST_PLAYBOOKS_DIR, WIDGETS_DIR,
-    XSOAR_AUTHOR, XSOAR_SUPPORT, XSOAR_SUPPORT_URL)
+    XSOAR_AUTHOR, XSOAR_SUPPORT, XSOAR_SUPPORT_URL, GithubContentConfig)
 from demisto_sdk.commands.common.tools import (LOG_COLORS,
                                                get_common_server_path,
                                                print_error, print_v,
@@ -209,6 +209,14 @@ class Initiator:
         elif os.path.isfile('content-descriptor.json'):
             self.full_output_path = os.path.join("Packs", self.dir_name)
 
+        # if in an external repo check for the existence of Packs directory
+        # if it does not exist create it
+        elif tools.is_external_repository():
+            if not os.path.isdir("Packs"):
+                print("Creating 'Packs' directory")
+                os.mkdir('Packs')
+            self.full_output_path = os.path.join("Packs", self.dir_name)
+
         # if non of the above conditions apply - create the pack in current directory
         else:
             self.full_output_path = self.dir_name
@@ -320,6 +328,10 @@ class Initiator:
             pack_metadata['email'] = input("\nThe email in which users can reach out for support (optional): ")
         else:  # community pack url should refer to the marketplace live discussions
             pack_metadata['url'] = MARKETPLACE_LIVE_DISCUSSIONS
+
+        dev_email = input("\nThe email will be used to inform you for any changes made to your pack (optional): ")
+        if dev_email:
+            pack_metadata['devEmail'] = [e.strip() for e in dev_email.split(',') if e]
 
         tags = input("\nTags of the pack, comma separated values: ")
         tags_list = [t.strip() for t in tags.split(',') if t]
@@ -571,7 +583,7 @@ class Initiator:
             bool. True if the files were downloaded and saved successfully, False otherwise.
         """
         # create test_data dir
-        if self.template in [self.HELLO_WORLD_INTEGRATION] + self.DEFAULT_TEMPLATES\
+        if self.template in [self.HELLO_WORLD_INTEGRATION] + self.DEFAULT_TEMPLATES \
                 + [self.HELLO_WORLD_FEED_INTEGRATION]:
             os.mkdir(os.path.join(self.full_output_path, self.TEST_DATA_DIR))
 
@@ -594,7 +606,12 @@ class Initiator:
                     # is `README_example.md` - which happens when we do not want the readme
                     # files to appear in https://xsoar.pan.dev/docs/reference/index.
                     filename = file.replace('README.md', 'README_example.md')
-                file_content = tools.get_remote_file(os.path.join(path, filename), return_content=True)
+                file_content = tools.get_remote_file(
+                    os.path.join(path, filename),
+                    return_content=True,
+                    # Templates available only in the official repo
+                    github_repo=GithubContentConfig.OFFICIAL_CONTENT_REPO_NAME
+                )
                 with open(os.path.join(self.full_output_path, file), 'wb') as f:
                     f.write(file_content)
             except Exception: